+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+------------
+
+### Added
+
+- Add a currency worksheet
+
+
 [0.8.1] - 2026-06-07
 --------------------
 
@@ -449,6 +457,7 @@ Phase 1: Core Domain.
 - Numeric literals with decimal point support
 - 373 passing tests
 
+[Unreleased]: https://github.com/wisnij/unitary/compare/v0.8.1...HEAD
 [0.8.1]: https://github.com/wisnij/unitary/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/wisnij/unitary/compare/v0.7.6...v0.8.0
 [0.7.6]: https://github.com/wisnij/unitary/compare/v0.7.5...v0.7.6